[API] Create OrganizationsToAttachToTargetProfileRepository
