+++ api/lib/domain/usecases/attach-organizations-to-target-profile.js
@@ -3,11 +3,11 @@ const { OrganizationsToAttachToTargetProfile } = require('../models');
 module.exports = async function attachOrganizationsToTargetProfile({
   targetProfileId,
   organizationIds,
-  targetProfileShareRepository,
+  organizationsToAttachToTargetProfileRepository,
 }) {
   const targetProfileOrganizations = new OrganizationsToAttachToTargetProfile({ id: targetProfileId });
 
   targetProfileOrganizations.attach(organizationIds);
 
-  return targetProfileShareRepository.attachOrganizations(targetProfileOrganizations);
+  return organizationsToAttachToTargetProfileRepository.attachOrganizations(targetProfileOrganizations);
 };
